+++ test/loop_c.cpp
@@ -8,7 +8,7 @@
 
 using namespace efp;
 
-constexpr size_t n = 10000;
+constexpr int n = 10000;
 
 int main(int argc, char const *argv[])
 {
@@ -20,7 +20,7 @@ int main(int argc, char const *argv[])
 
     std::iota(std::begin(xs), std::end(xs), 0);
 
-    // for (int i = 0; i < n; ++i)
+    // for (size_t i = 0; i < n; ++i)
     // {
     //     xs[i] = std::rand();
     // }
@@ -28,54 +28,54 @@ int main(int argc, char const *argv[])
     // Vector<int> ys(n);
     Array<int, n> ys{};
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = square(xs[i]);
     }
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = sqrt<double>(ys[i]);
     }
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = square<double>(ys[i]);
     }
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = sqrt<double>(ys[i]);
     }
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = square<double>(ys[i]);
     }
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = sqrt<double>(ys[i]);
     }
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = square<double>(ys[i]);
     }
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = sqrt<double>(ys[i]);
     }
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = square<double>(ys[i]);
     }
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         ys[i] = sqrt<double>(ys[i]);
     }
 
     double sum = 0;
 
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < n; ++i)
     {
         sum += ys[i];
     }
